refactor(admin): extract AuditTab into its own file
